feat: 添加概率直通和轻量 LLM 判定处理器，优化预处理事件逻辑
