+++ plugins/neo_default_chatter/components/event_handlers/defaults/preprocess/probability_bypass.py
@@ -21,12 +21,12 @@ from src.app.plugin_system.types import ChatStream, Message
 from src.core.config import get_core_config
 from src.kernel.event import EventDecision
 
-from ..config import NeoChatterConfig
-from ...utils.event_publisher import NdfcEvent
+from ....config import NeoChatterConfig
+from .....utils.event_publisher import NdfcEvent
 
 logger = get_logger("neo_default_chatter.preprocess.probability_bypass")
 
-#: NFC 预处理事件名。与 ``utils.preprocess.PREPROCESS_EVENT`` 同源。
+#: NFC 预处理事件名。
 _PREPROCESS_EVENT = NdfcEvent.PREPROCESS
